+++ lib/flutter_shogi_board.dart
@@ -1,3 +1,4 @@
 library flutter_shogi_board;
 
 export 'src/widgets/shogi_board.dart';
+export 'src/widgets/coord_indicator_cell.dart' show CoordIndicatorType;