+++ lib/src/widgets/shogi_board.dart
@@ -4,6 +4,7 @@ import 'package:shogi/shogi.dart';
 import '../configs/board_colors.dart';
 import '../utils/package_utils.dart';
 import '../widgets/board_cell.dart';
+import '../widgets/coord_indicator_cell.dart';
 
 /// Renders a shogi board using a list of board pieces
 class ShogiBoard extends StatelessWidget {
@@ -22,8 +23,14 @@ class ShogiBoard extends StatelessWidget {
   /// The color of each cell's border
   final Color borderColor;
 
-  /// If `true` uses japanese characters (i.e. 玉), otherwise english letters (i.e. K)
+  /// Whether japanese characters (i.e. 玉) or latin letters (i.e. K) should be used. Defaults to `true`.
   final bool usesJapanese;
+
+  /// Whether coordinate indicators (on top and right of board) should be shown. Defaults to `true`.
+  final bool showCoordIndicators;
+
+  /// The type of coordinate indicators show. Defaults to `CoordIndicatorType.japanese`.
+  final CoordIndicatorType coordIndicatorType;
 
   const ShogiBoard({
     @required this.boardPieces,
@@ -32,35 +39,47 @@ class ShogiBoard extends StatelessWidget {
     this.cellColor = Colors.transparent,
     this.borderColor = BoardColors.gray,
     this.usesJapanese = true,
+    this.showCoordIndicators = true,
+    this.coordIndicatorType = CoordIndicatorType.japanese,
     Key key,
   }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
+    final numberRows = showCoordIndicators ? BoardConfig.numberRows + 1 : BoardConfig.numberRows;
+    final numberColumns = showCoordIndicators ? BoardConfig.numberColumns + 1 : BoardConfig.numberColumns;
+
     return LayoutBuilder(
       builder: (_, constraints) {
-        final size =
-            (constraints.hasBoundedWidth ? constraints.maxWidth : constraints.maxHeight) / BoardConfig.numberRows;
+        final size = (constraints.hasBoundedWidth ? constraints.maxWidth : constraints.maxHeight) / numberRows;
 
-        List<Widget> rows = List<Widget>(BoardConfig.numberRows);
-        for (int y = 0; y < BoardConfig.numberRows; y++) {
-          List<Widget> row = List<Widget>(BoardConfig.numberColumns);
-          for (int x = BoardConfig.numberColumns - 1; x >= 0; x--) {
-            final boardPiece = PackageUtils.pieceAtPosition(boardPieces, x, y);
-            row[BoardConfig.numberColumns - 1 - x] = BoardCell(
-              boardPiece: boardPiece?.displayString(usesJapanese: usesJapanese) ?? '',
-              sente: boardPiece?.isSente ?? true,
-              size: size,
-              edge: Edge(
-                top: y == 0,
-                bottom: y == BoardConfig.numberRows - 1,
-                left: x == BoardConfig.numberColumns - 1,
-                right: x == 0,
-              ),
-              pieceColor: boardPiece != null ? (boardPiece.isPromoted ? promotedPieceColor : pieceColor) : null,
-              cellColor: cellColor,
-              borderColor: borderColor,
+        List<Widget> rows = List<Widget>(numberRows);
+        for (int y = 0; y < numberRows; y++) {
+          List<Widget> row = List<Widget>(numberColumns);
+          for (int x = numberColumns - 1; x >= 0; x--) {
+            final boardPiece = PackageUtils.pieceAtPosition(
+              boardPieces,
+              showCoordIndicators ? y : y + 1,
+              showCoordIndicators ? x : x + 1,
             );
+
+            row[numberColumns - 1 - x] = showCoordIndicators && (y == 0 || x == 0)
+                ? CoordIndicatorCell(
+                    size: size, coord: y == 0 ? x : y, isTop: y == 0, coordIndicatorType: coordIndicatorType)
+                : BoardCell(
+                    boardPiece: boardPiece?.displayString(usesJapanese: usesJapanese) ?? '',
+                    sente: boardPiece?.isSente ?? true,
+                    size: size,
+                    edge: Edge(
+                      top: y == (showCoordIndicators ? 1 : 0),
+                      bottom: y == numberRows - 1,
+                      left: x == numberColumns - 1,
+                      right: x == (showCoordIndicators ? 1 : 0),
+                    ),
+                    pieceColor: boardPiece != null ? (boardPiece.isPromoted ? promotedPieceColor : pieceColor) : null,
+                    cellColor: cellColor,
+                    borderColor: borderColor,
+                  );
           }
           rows[y] = Row(
             mainAxisAlignment: MainAxisAlignment.spaceEvenly,